+++ docs/CHANGELOG.md
@@ -4,4 +4,5 @@
 - Fix getTable method when the user set a subquery
 eg h2GIS.getSpatialTable("h2gis").columns("land", "st_precisionreducer(st_transform(the_geom, 4326), 3) as the_geom")
   .filter("limit 1").getSpatialTable()
+- Add methods on datasource to create, drop indexes, drop table, drop columns
 - Update H2GIS to 2.2.1-SNAPSHOT